refactor(util): give CDN version resolution its own module
The last half of #293's suggested shape. `maidr/util/cdn.py` takes the
resolver -- the endpoints, the timeout budget, the semver pin handling,
the `latest` lookup and its caching, and the URL builders.
`dependencies.py` goes 1781 -> 835 lines and is finally what the issue
asked for: reaching the copy bundled inside the wheel.

The direction is one-way. Measured before moving anything rather than
assumed: no asset-side function references a CDN name, so `cdn` imports
`dependencies` and nothing imports back.

Two names stayed behind for the reason #293 records after the first cut.
`_version_key` and `_UNKNOWN_VERSION` are used by `bundle_freshness`, so
they are shared version primitives rather than CDN ones, and moving them
would have made freshness depend on the resolver to compare two local
strings. `_warn_placeholder_css` stayed for the same reason found later:
`bundled_css_path` calls it too, and it is about the placeholder asset,
not the CDN.

The back-compat shim is weaker here than for the earlier cuts, and the
difference is worth stating rather than discovering. It forwards
attribute **reads**, so `dependencies.get_cdn_version` still resolves to
the same object; it cannot forward a **write**, because `setattr` goes
straight to the module dict. So
`monkeypatch.setattr(dependencies, "urlopen", ...)` would set an
attribute nothing in `cdn` reads -- a patch that silently does nothing.
That is why only the public names are forwarded, why the resolver's
state is deliberately absent from `_MOVED_TO_CDN`, and why 266 test
references were repointed at `cdn` instead of being left to the shim.
`test_the_shim_forwards_reads_but_cannot_forward_a_patch` asserts the
asymmetry so a future change that starts forwarding writes has to edit
it.

Two tests needed real updates, both predicted:

- `test_no_render_path_references_the_unresolved_constants` is the
  exclusion #293's third note said would need repointing once another
  module owned `MAIDR_JS_CDN_URL` / `MAIDR_CSS_CDN_URL`. It now allows
  the owner *and* the shim's forwarding list, which is not a render path.
- `maidr_css_cdn_url`'s deprecation named
  `cdn_url(MAIDR_MATH_CSS_FILENAME) from maidr.util.dependencies`, which
  after the move sent readers to a module where neither name lives.
  `test_deprecation_names_only_importable_symbols` caught it; the advice
  now names `cdn`, and the filename is re-exported there so both halves
  resolve where the reader is sent.

The laziness guard covers `cdn` too. For this pair every module-scope
import spelling fails loudly rather than quietly -- `cdn` from-imports
`dependencies`, so the cycle raises `partially initialized module` --
but the guard is not vacuous: checked directly, it flags
`import maidr.util.cdn`, `from maidr.util import cdn` and
`from maidr.util.cdn import cdn_url` alike.
